+++ docusaurus/website/sidebars.js
@@ -58,6 +58,7 @@ const sidebars = {
         'advanced-behaviors',
         'advanced-custom-datasource',
         'advanced-time-range-comparison',
+        'url-sync',
       ],
     },
     {
@@ -65,13 +66,10 @@ const sidebars = {
       label: '@grafana/scenes-ml',
       collapsible: true,
       collapsed: false,
-      items: [
-        'getting-started',
-        'baselines-and-forecasts',
-        'outlier-detection',
-        'changepoint-detection',
-      ].map(id => `scenes-ml/${id}`),
-    }
+      items: ['getting-started', 'baselines-and-forecasts', 'outlier-detection', 'changepoint-detection'].map(
+        (id) => `scenes-ml/${id}`
+      ),
+    },
   ],
 };
 module.exports = sidebars;