+++ packages/scenes/src/core/SceneTimeRange.tsx
@@ -1,4 +1,4 @@
-import { getTimeZone, rangeUtil, setWeekStart, TimeRange } from '@grafana/data';
+import { getTimeZone, rangeUtil, setWeekStart, TimeRange, toUtc } from '@grafana/data';
 import { TimeZone } from '@grafana/schema';
 
 import { SceneObjectUrlSyncConfig } from '../services/SceneObjectUrlSyncConfig';
@@ -8,10 +8,10 @@ import { SceneTimeRangeLike, SceneTimeRangeState, SceneObjectUrlValues } from '.
 import { getClosest } from './sceneGraph/utils';
 import { parseUrlParam } from '../utils/parseUrlParam';
 import { evaluateTimeRange } from '../utils/evaluateTimeRange';
-import { config, RefreshEvent } from '@grafana/runtime';
+import { config, locationService, RefreshEvent } from '@grafana/runtime';
 
 export class SceneTimeRange extends SceneObjectBase<SceneTimeRangeState> implements SceneTimeRangeLike {
-  protected _urlSync = new SceneObjectUrlSyncConfig(this, { keys: ['from', 'to', 'timezone'] });
+  protected _urlSync = new SceneObjectUrlSyncConfig(this, { keys: ['from', 'to', 'timezone', 'time', 'time.window'] });
 
   public constructor(state: Partial<SceneTimeRangeState> = {}) {
     const from = state.from ?? 'now-6h';
@@ -23,9 +23,8 @@ export class SceneTimeRange extends SceneObjectBase<SceneTimeRangeState> impleme
       timeZone || getTimeZone(),
       state.fiscalYearStartMonth,
       state.UNSAFE_nowDelay
-
     );
-    const refreshOnActivate = state.refreshOnActivate ?? {percent: 10}
+    const refreshOnActivate = state.refreshOnActivate ?? { percent: 10 };
     super({ from, to, timeZone, value, refreshOnActivate, ...state });
 
     this.addActivationHandler(this._onActivate.bind(this));
@@ -58,7 +57,7 @@ export class SceneTimeRange extends SceneObjectBase<SceneTimeRangeState> impleme
       setWeekStart(this.state.weekStart);
     }
 
-    if(rangeUtil.isRelativeTimeRange(this.state.value.raw)){
+    if (rangeUtil.isRelativeTimeRange(this.state.value.raw)) {
       this.refreshIfStale();
     }
 
@@ -120,16 +119,16 @@ export class SceneTimeRange extends SceneObjectBase<SceneTimeRangeState> impleme
     );
 
     const diff = value.to.diff(this.state.value.to, 'milliseconds');
-    if(diff >= refreshAfterMs){
+    if (diff >= refreshAfterMs) {
       this.setState({
-        value
-      })
+        value,
+      });
     }
   }
 
   private calculatePercentOfInterval(percent: number): number {
     const intervalMs = this.state.value.to.diff(this.state.value.from, 'milliseconds');
-    return Math.ceil(intervalMs / percent)
+    return Math.ceil(intervalMs / percent);
   }
 
   public getTimeZone(): TimeZone {
@@ -178,12 +177,16 @@ export class SceneTimeRange extends SceneObjectBase<SceneTimeRangeState> impleme
 
     // Only update if time range actually changed
     if (update.from !== this.state.from || update.to !== this.state.to) {
-      this.setState(update);
+      this._urlSync.performBrowserHistoryAction(() => {
+        this.setState(update);
+      });
     }
   };
 
   public onTimeZoneChange = (timeZone: TimeZone) => {
-    this.setState({ timeZone });
+    this._urlSync.performBrowserHistoryAction(() => {
+      this.setState({ timeZone });
+    });
   };
 
   public onRefresh = () => {
@@ -197,32 +200,48 @@ export class SceneTimeRange extends SceneObjectBase<SceneTimeRangeState> impleme
       ),
     });
 
-    this.publishEvent(new RefreshEvent());
+    this.publishEvent(new RefreshEvent(), true);
   };
 
   public getUrlState() {
+    const params = locationService.getSearchObject();
     const urlValues: SceneObjectUrlValues = { from: this.state.from, to: this.state.to };
+
     if (this.state.timeZone) {
       urlValues.timezone = this.state.timeZone;
+    }
+
+    // Clear time and time.window once they are converted to from and to
+    if (params.time && params['time.window']) {
+      urlValues.time = null;
+      urlValues['time.window'] = null;
     }
 
     return urlValues;
   }
 
   public updateFromUrl(values: SceneObjectUrlValues) {
-    // ignore if both are missing
-    if (!values.to && !values.from) {
-      return;
+    const update: Partial<SceneTimeRangeState> = {};
+
+    let from = parseUrlParam(values.from);
+    let to = parseUrlParam(values.to);
+
+    if (values.time && values['time.window']) {
+      const time = Array.isArray(values.time) ? values.time[0] : values.time;
+      const timeWindow = Array.isArray(values['time.window']) ? values['time.window'][0] : values['time.window'];
+      const timeRange = getTimeWindow(time, timeWindow);
+      from = timeRange.from;
+      to = timeRange.to;
     }
 
-    const update: Partial<SceneTimeRangeState> = {};
-    const from = parseUrlParam(values.from);
+    if (!from && !to) {
+      return;
+    }
 
     if (from) {
       update.from = from;
     }
 
-    const to = parseUrlParam(values.to);
     if (to) {
       update.to = to;
     }
@@ -241,4 +260,28 @@ export class SceneTimeRange extends SceneObjectBase<SceneTimeRangeState> impleme
 
     this.setState(update);
   }
+}
+
+/**
+ * Calculates the duration of the time range from time-time.window/2 to time+time.window/2. Both be specified in ms. For example ?time=1500000000000&time.window=10000 results in a 10-second time range from 1499999995000 to 1500000005000`.
+ * @param time - time in ms
+ * @param timeWindow - time window in ms or interval string
+ */
+function getTimeWindow(time: string, timeWindow: string) {
+  // Parse the time, assuming it could be an ISO string or a number in milliseconds
+  const valueTime = isNaN(Date.parse(time)) ? parseInt(time, 10) : Date.parse(time);
+
+  let timeWindowMs;
+
+  if (timeWindow.match(/^\d+$/) && parseInt(timeWindow, 10)) {
+    // when time window is specified in ms
+    timeWindowMs = parseInt(timeWindow, 10);
+  } else {
+    timeWindowMs = rangeUtil.intervalToMs(timeWindow);
+  }
+
+  return {
+    from: toUtc(valueTime - timeWindowMs / 2).toISOString(),
+    to: toUtc(valueTime + timeWindowMs / 2).toISOString(),
+  };
 }