+++ packages/scenes/src/variables/adhoc/AdHocFiltersVariable.tsx
@@ -15,12 +15,15 @@ import { getDataSourceSrv } from '@grafana/runtime';
 import { AdHocFiltersVariableUrlSyncHandler } from './AdHocFiltersVariableUrlSyncHandler';
 import { css } from '@emotion/css';
 import { getEnrichedFiltersRequest } from '../getEnrichedFiltersRequest';
-import { SafeSerializableSceneObject } from '../../utils/SafeSerializableSceneObject';
+import { AdHocFiltersComboboxRenderer } from './AdHocFiltersCombobox/AdHocFiltersComboboxRenderer';
+import { wrapInSafeSerializableSceneObject } from '../../utils/wrapInSafeSerializableSceneObject';
 
 export interface AdHocFilterWithLabels extends AdHocVariableFilter {
   keyLabel?: string;
-  valueLabel?: string;
+  valueLabels?: string[];
 }
+
+export type AdHocControlsLayout = ControlsLayout | 'combobox';
 
 export interface AdHocFiltersVariableState extends SceneVariableState {
   /** Optional text to display on the 'add filter' button */
@@ -37,7 +40,7 @@ export interface AdHocFiltersVariableState extends SceneVariableState {
    * @experimental
    * Controls the layout and design of the label.
    */
-  layout?: ControlsLayout;
+  layout?: AdHocControlsLayout;
   /**
    * Defaults to automatic which means filters will automatically be applied to all queries with the same data source as this AdHocFilterSet.
    * In manual mode you either have to use the filters programmatically or as a variable inside query expressions.
@@ -78,6 +81,11 @@ export interface AdHocFiltersVariableState extends SceneVariableState {
   expressionBuilder?: AdHocVariableExpressionBuilderFn;
 
   /**
+   * Whether the filter supports new multi-value operators like =| and !=|
+   */
+  supportsMultiValueOperators?: boolean;
+
+  /**
    * When querying the datasource for label names and values to determine keys and values
    * for this ad hoc filter, consider the queries in the scene and use them as a filter.
    * This queries filter can be used to ensure that only ad hoc filter options that would
@@ -105,13 +113,52 @@ export type getTagValuesProvider = (
 
 export type AdHocFiltersVariableCreateHelperArgs = AdHocFiltersVariableState;
 
+export type OperatorDefinition = {
+  value: string;
+  description?: string;
+  isMulti?: Boolean;
+};
+
+export const OPERATORS: OperatorDefinition[] = [
+  {
+    value: '=',
+  },
+  {
+    value: '!=',
+  },
+  {
+    value: '=|',
+    description: 'Is one of. Use to filter on multiple values.',
+    isMulti: true,
+  },
+  {
+    value: '!=|',
+    description: 'Is not one of. Use to exclude multiple values.',
+    isMulti: true,
+  },
+  {
+    value: '=~',
+    description: 'Matches regex',
+  },
+  {
+    value: '!~',
+    description: 'Does not match regex',
+  },
+  {
+    value: '<',
+  },
+  {
+    value: '>',
+  },
+];
+
 export class AdHocFiltersVariable
   extends SceneObjectBase<AdHocFiltersVariableState>
   implements SceneVariable<AdHocFiltersVariableState>
 {
   static Component = AdHocFiltersVariableRenderer;
 
-  private _scopedVars = { __sceneObject: new SafeSerializableSceneObject(this) };
+  private _scopedVars = { __sceneObject: wrapInSafeSerializableSceneObject(this) };
   private _dataSourceSrv = getDataSourceSrv();
 
   protected _urlSync = new AdHocFiltersVariableUrlSyncHandler(this);
@@ -151,41 +198,21 @@ export class AdHocFiltersVariable
     return this.state.filterExpression;
   }
 
-  public _updateFilter(
-    filter: AdHocFilterWithLabels,
-    prop: keyof AdHocFilterWithLabels,
-    { value, label }: SelectableValue<string | undefined | null>
-  ) {
-    if (value == null) {
-      return;
-    }
-
+  public _updateFilter(filter: AdHocFilterWithLabels, update: Partial<AdHocFilterWithLabels>) {
     const { filters, _wip } = this.state;
-
-    const propLabelKey = `${prop}Label`;
 
     if (filter === _wip) {
       // If we set value we are done with this "work in progress" filter and we can add it
-      if (prop === 'value') {
-        this.setState({ filters: [...filters, { ..._wip, [prop]: value, [propLabelKey]: label }], _wip: undefined });
+      if ('value' in update && update['value'] !== '') {
+        this.setState({ filters: [...filters, { ..._wip, ...update }], _wip: undefined });
       } else {
-        this.setState({ _wip: { ...filter, [prop]: value, [propLabelKey]: label } });
+        this.setState({ _wip: { ...filter, ...update } });
       }
       return;
     }
 
     const updatedFilters = this.state.filters.map((f) => {
-      if (f === filter) {
-        const updatedFilter = { ...f, [prop]: value, [propLabelKey]: label };
-
-        // clear value if key has changed
-        if (prop === 'key' && filter[prop] !== value) {
-          updatedFilter.value = '';
-          updatedFilter.valueLabel = '';
-        }
-        return updatedFilter;
-      }
-      return f;
+      return f === filter ? { ...f, ...update } : f;
     });
 
     this.setState({ filters: updatedFilters });
@@ -198,6 +225,14 @@ export class AdHocFiltersVariable
     }
 
     this.setState({ filters: this.state.filters.filter((f) => f !== filter) });
+  }
+
+  public _removeLastFilter() {
+    const filterToRemove = this.state.filters.at(-1);
+
+    if (filterToRemove) {
+      this._removeFilter(filterToRemove);
+    }
   }
 
   /**
@@ -293,14 +328,18 @@ export class AdHocFiltersVariable
 
   public _addWip() {
     this.setState({
-      _wip: { key: '', keyLabel: '', value: '', valueLabel: '', operator: '=', condition: '' },
+      _wip: { key: '', value: '', operator: '=', condition: '' },
     });
   }
 
   public _getOperators() {
-    return ['=', '!=', '<', '>', '=~', '!~'].map<SelectableValue<string>>((value) => ({
+    const filteredOperators = this.state.supportsMultiValueOperators
+      ? OPERATORS
+      : OPERATORS.filter((operator) => !operator.isMulti);
+    return filteredOperators.map<SelectableValue<string>>(({ value, description }) => ({
       label: value,
       value,
+      description,
     }));
   }
 }
@@ -315,6 +354,10 @@ function renderExpression(
 export function AdHocFiltersVariableRenderer({ model }: SceneComponentProps<AdHocFiltersVariable>) {
   const { filters, readOnly, addFilterButtonText } = model.useState();
   const styles = useStyles2(getStyles);
+
+  if (model.state.layout === 'combobox') {
+    return <AdHocFiltersComboboxRenderer model={model} />;
+  }
 
   return (
     <div className={styles.wrapper}>
@@ -337,10 +380,6 @@ const getStyles = (theme: GrafanaTheme2) => ({
     columnGap: theme.spacing(2),
     rowGap: theme.spacing(1),
   }),
-  filterIcon: css({
-    color: theme.colors.text.secondary,
-    paddingRight: theme.spacing(0.5),
-  }),
 });
 
 export function toSelectableValue(input: MetricFindValue): SelectableValue<string> {
@@ -359,4 +398,12 @@ export function toSelectableValue(input: MetricFindValue): SelectableValue<strin
 
 export function isFilterComplete(filter: AdHocFilterWithLabels): boolean {
   return filter.key !== '' && filter.operator !== '' && filter.value !== '';
+}
+
+export function isMultiValueOperator(operatorValue: string): boolean {
+  const operator = OPERATORS.find((o) => o.value === operatorValue);
+  if (!operator) {
+    throw new Error('Unknown operator');
+  }
+  return Boolean(operator.isMulti);
 }